req3_llm_rag/requirements.txt
# Retrieval
torch>=2.1.0
sentence-transformers>=2.7.0
qdrant-client>=1.8.0

# LLM
aiohttp>=3.9.0

# Utils
python-dotenv>=1.0.0
rich>=13.7.0
scikit-learn>=1.4.0
numpy>=1.26.0

# API
fastapi>=0.110.0
uvicorn>=0.27.0
